changed names of few JMenuItems to reflect what they actually are.
added functionality to the Start New Game menu item. It will now reset the score, paddles, and ball for the players.
added a getScore method to Game.java to achieve this and utilize the Score.java reset() method. Then utilized drawNewBall() and drawNewPaddles() used to start the game to reset the ball and paddles.