Added singleplayer and multiplayer functionality
Added a new property property to the game class; a boolean called multiplayer which by default is true. If the game is set to multiplayer then it will move player2 (right paddle) by using the arrow keys. If set to singleplayer then player2 will be acting as an AI which will follow the ball at a velocity that is 1 less than that of the balls velocity so it has a chance of catching up but it wont ALWAYS catch the ball.

When the menu is clicked it will change the value of multiplayer by calling respectively either setSinglePlayer or setMultiplayer depending on which option from the menu is clicked.